Allow to function in vscode web
