+++ .vscodeignore
@@ -1,5 +1,6 @@
 .vscode/**
 .vscode-test/**
+.vscode-test-web/**
 node_modules/**
 coverage/**
 out/test/**
