+++ package.json
@@ -114,12 +114,14 @@
     "watch": "tsc -watch -p ./",
     "pretest": "tsc -p ./",
     "test": "node ./out/test/runTest.js",
+    "web": "npm run build && vscode-test-web --browserType=webkit --extensionDevelopmentPath=./",
     "esbuild-base": "esbuild ./src/extension.ts --bundle --outfile=out/src/extension.js --external:vscode --format=cjs --platform=node"
   },
   "devDependencies": {
     "@types/mocha": "^5.2.6",
     "@types/node": "^6.14.13",
     "@types/vscode": "^1.54.0",
+    "@vscode/test-web": "^0.0.7",
     "esbuild": "^0.12.19",
     "coveralls": "^3.1.1",
     "mocha": "^9.0.3",
